+++ subgraphs/arbitrum/src/treasury/Assets.ts
@@ -6,8 +6,8 @@ import {
   TokenCategoryVolatile,
 } from "../../../shared/src/contracts/TokenDefinition";
 import { pushTokenRecordArray, pushTokenSupplyArray } from "../../../shared/src/utils/ArrayHelper";
-import { FundsDeposited } from "../../generated/TokenRecords-arbitrum/GelatoTaskTreasury";
-import { getProtocolOwnedLiquiditySupplyRecords, getTreasuryOHMRecords } from "./OhmCalculations";
+import { NewRound } from "../../generated/TokenRecords-arbitrum/ChainlinkAggregator";
+import { getProtocolOwnedLiquiditySupplyRecords, getTotalSupply, getTreasuryOHMRecords } from "./OhmCalculations";
 import { getOwnedLiquidityBalances } from "./OwnedLiquidity";
 import { getTokenBalances } from "./TokenBalances";
 
@@ -36,6 +36,12 @@ function generateTokenRecords(timestamp: BigInt, blockNumber: BigInt): TokenReco
 function generateTokenSupplies(timestamp: BigInt, blockNumber: BigInt): TokenSupply[] {
   const records: TokenSupply[] = [];
 
+  // Total supply
+  pushTokenSupplyArray(
+    records,
+    getTotalSupply(timestamp, blockNumber),
+  );
+
   // Treasury OHM
   pushTokenSupplyArray(
     records,
@@ -51,7 +57,7 @@ function generateTokenSupplies(timestamp: BigInt, blockNumber: BigInt): TokenSup
   return records;
 }
 
-export function handleEvent(event: FundsDeposited): void {
+export function handleEvent(event: NewRound): void {
   const block = event.block;
 
   log.debug("handleEvent: *** Indexing block {}", [block.number.toString()]);