+++ subgraphs/arbitrum/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Arbitrum Subgraph Changelog
 
+## 1.3.0 (2023-05-XX)
+
+- Add LUSD, LQTY on Arbitrum
+- Add OHM on Arbitrum
+- Add Chainlink price feeds for LUSD, USDC, WETH
+- Add OHM total supply
+
 ## 1.2.2 (2023-04-29)
 
 - Populate TokenSupply records with the quantity of protocol-owned gOHM